Move binary_output_style proto <-> enum mapping to core
