+++ native/spark-expr/Cargo.toml
@@ -37,8 +37,6 @@ thiserror = { workspace = true }
 futures = { workspace = true }
 twox-hash = "2.1.2"
 rand = { workspace = true }
-datafusion-comet-proto = { workspace = true }
-prost = "0.14.1"
 hex = "0.4.3"
 base64 = "0.22.1"
